refactor: move validators to outer classes
