+++ src/ak/po/Entities/Product.java
@@ -1,7 +1,10 @@
 package ak.po.Entities;
 
-import ak.po.Entities.Category;
 import ak.po.Enums.Currency;
+import ak.po.Validators.CategoryValidator;
+import ak.po.Validators.CurrencyValidator;
+import ak.po.Validators.FloatValidator;
+import ak.po.Validators.StringValidator;
 
 public class Product {
     private final String name;
@@ -23,113 +26,31 @@ public class Product {
         private Currency currency = null;
 
         public Builder name(String name) {
-            ProductValidator.getInstance().validate(name);
+            StringValidator.getInstance().validate(name);
             this.name = name;
             return this;
         }
 
         public Builder category(Category category) {
-            ProductValidator.getInstance().validate(category);
+            CategoryValidator.getInstance().validate(category);
             this.category = category;
             return this;
         }
 
         public Builder price(float price) {
-            ProductValidator.getInstance().validate(price);
+            FloatValidator.getInstance().validate(price);
             this.price = price;
             return this;
         }
 
         public Builder currency(Currency currency) {
-            ProductValidator.getInstance().validate(currency);
+            CurrencyValidator.getInstance().validate(currency);
             this.currency = currency;
             return this;
         }
 
         public Product build() {
             return new Product(this);
-        }
-
-        public static class ProductValidator {
-            private final static ProductValidator instance = new ProductValidator();
-
-            public static ProductValidator getInstance() {
-                return instance;
-            }
-
-            public static boolean validate(String name) {
-                if(!nullValidator(name)) {
-                    return false;
-                } else if(!lengthValidator(name)) {
-                    return false;
-                } else {
-                    return true;
-                }
-            }
-
-            public static boolean validate(Category category) {
-                if(!nullValidator(category)) {
-                    return false;
-                } else {
-                    return true;
-                }
-            }
-
-            public static boolean validate(float price) {
-                if(!negativeValueValidator(price)) {
-                    return false;
-                } else {
-                    return true;
-                }
-            }
-
-            public static boolean validate(Currency currency) {
-                if(!nullValidator(currency)) {
-                    return false;
-                } else {
-                    return true;
-                }
-            }
-
-            public static boolean lengthValidator(String name) {
-                if(name.isEmpty()) {
-                    throw new IllegalArgumentException("Argument nie może być pusty");
-                } else {
-                    return true;
-                }
-            }
-
-            public static boolean nullValidator(String name) {
-                if(name == null) {
-                    throw new IllegalArgumentException("Argument nie może mieć wartości NULL");
-                } else {
-                    return true;
-                }
-            }
-
-            public static boolean nullValidator(Category category) {
-                if(category == null) {
-                    throw new IllegalArgumentException("Argument nie może mieć wartości NULL");
-                } else {
-                    return true;
-                }
-            }
-
-            public static boolean nullValidator(Currency currency) {
-                if(currency == null) {
-                    throw new IllegalArgumentException("Argument nie może mieć wartości NULL");
-                } else {
-                    return true;
-                }
-            }
-
-            public static boolean negativeValueValidator(float price) {
-                if(price < 0) {
-                    throw new IllegalArgumentException("Cena musi być większa od zera");
-                } else {
-                    return true;
-                }
-            }
         }
     }
